% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{aid2r.aid}
\alias{aid2r.aid}
\title{Convert aid to R aid}
\usage{
aid2r.aid(x)
}
\arguments{
\item{x}{A vector of AIDs}
}
\description{
This function converts aid into R aid
}
\examples{
aid2r.aid()
}
\keyword{aid}
\keyword{r.aid}
