% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer_info.R
\name{cancer2code}
\alias{cancer2code}
\title{Get TCGA code by cancer name}
\usage{
cancer2code(cancer)
}
\arguments{
\item{cancer}{the Chinese cancer name}
}
\value{
TCGA codes
}
\description{
Get TCGA code by cancer name
}
\examples{
\dontrun{
cancer2code("胃癌")
}
}
