% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer_info}
\alias{cancer_info}
\title{TCGA cancer code}
\format{
A data frame with 3 variables: \code{code}, \code{cancer} and
\code{cancer_name_cn}
}
\usage{
cancer_info
}
\description{
words frequency from sina news and yk.
}
\keyword{datasets}
