% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csv_cols}
\alias{csv_cols}
\title{The columns need read}
\format{
A named vector. columns: \code{names(csv_cols)},
type: \code{csv_cols}.
}
\usage{
csv_cols
}
\description{
The default columns to read in .csv files.
}
\keyword{datasets}
