% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Format data for next step}
\usage{
format_data(mua, db_path, sample_info, match = FALSE)
}
\arguments{
\item{mua}{mutation data.frame}

\item{db_path}{db path}

\item{sample_info}{sample info data.frame}
}
\value{
data.table
}
\description{
Format data for next step
}
\examples{
\dontrun{
res <- format_data(mua, db_path, sample_info)
}
}
