% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{func.anova}
\alias{func.anova}
\title{Get AOV pvalue}
\usage{
func.anova(x, g)
}
\arguments{
\item{x}{Continuous data vector}

\item{g}{Group labels}
}
\description{
This function returns the -log10(pvalue) of AOV
}
\examples{
func.anova(x=c(3,4,5,10,12,15,100,120,130), g=c("A","A","A","B","B","B","C","C","C"))
}
\keyword{aov}
