% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{get_ratio}
\alias{get_ratio}
\title{Count the ratio}
\usage{
get_ratio(mutations, na = FALSE, stages = TRUE, asian = FALSE)
}
\arguments{
\item{mutations}{form mutations}

\item{na}{filter na}

\item{asian}{filer asian}
}
\value{
data.table
}
\description{
Count the ratio
}
\examples{
\dontrun{
get_ratio(mutations)
}
}
