% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer_info.R
\name{get_result}
\alias{get_result}
\title{Mutation}
\usage{
get_result(cancer, tcga_path, db_path)
}
\arguments{
\item{cancer}{The Chinese cancer name}

\item{tcga_path}{The path to TCGA DB}

\item{db_path}{THE path to DB}
}
\value{

}
\description{
Mutation
}
\examples{
\dontrun{
get_result(cancer, tcga_path,  db_path)
}
}
