% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{get_sample_info}
\alias{get_sample_info}
\title{Get the sample info}
\usage{
get_sample_info(data_path, index, bind = FALSE)
}
\arguments{
\item{data_path}{File name in working directory, path to file that contain
sample id, stage...}

\item{index}{A vector of column names to keep. Just look like
c(sample_id = col_index, race = col_index2, stages = col_index3). Index can be
column name or the number index}

\item{bind}{FALSE useless}
}
\value{
sample info
}
\description{
Get the sample info
}
\examples{
\dontrun{
index <- c(sample_id = "submitter_id", race = "race", stages = "tumor_stage")
sample_info <- get_sample_info("data_path", index)
}
}
