% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maf_cols}
\alias{maf_cols}
\title{The columns need read}
\format{
A named vector. columns: \code{names(maf_cols)},
type: \code{maf_cols}.
}
\usage{
maf_cols
}
\description{
The default columns to read in .maf files.
}
\keyword{datasets}
