% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{mutation}
\alias{mutation}
\title{Count mutation}
\usage{
mutation(df, stage = FALSE)
}
\arguments{
\item{df}{from get_result()}

\item{stage}{filter NA}
}
\value{

}
\description{
Count mutation
}
\examples{
\dontrun{
mutation(get_result(cancer, tcga_path, db_path))
}
}
