% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{r.aid2aid}
\alias{r.aid2aid}
\title{Convert R aid to aid}
\usage{
r.aid2aid(x)
}
\arguments{
\item{x}{A vector of R AIDs.}
}
\description{
This function converts R AIDs into AIDs
}
\examples{
r.aid2aid()
}
\keyword{aid}
\keyword{r.aid}
