% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ratio_plot}
\alias{ratio_plot}
\title{Plot the ratio}
\usage{
ratio_plot(out)
}
\arguments{
\item{mutations}{the data.frame}

\item{asian}{filter asian}

\item{na}{omit.na}
}
\value{
plot
}
\description{
Plot the ratio
}
\examples{
\dontrun{
ratio_plot(df)
}
}
