% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{sampleTypeToFactor}
\alias{sampleTypeToFactor}
\title{Sample types to Factors}
\usage{
sampleTypeToFactor(v)
}
\arguments{
\item{v}{Vector of Sample types}
}
\description{
This function converts numerical TCGA Sample types to a factor by replacing with "T" or "N"
}
\examples{
sampleTypeToFactor(c(1,2,3,6,10,11))
}
\keyword{sample}
\keyword{type}
