\name{tcga-package}
\alias{tcga-package}
\alias{tcga}
\docType{package}
\title{Package to reformat TCGA data into Sage Bionetworks format}
\description{
This package contains functions to reformat data downloaded from the TCGA
project into the Sage Bionetworks format.  It insures ID consistency across
data types and can package data in either RData or text files.
}
\details{
\tabular{ll}{
Package: \tab tcga\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-11-10\cr
License: \tab Apache Software Foundation 2.0\cr
LazyLoad: \tab yes\cr
}
}
\author{
Justin Guinney and David Henderson.
Maintainer: Sage Bionetworks Data Repository Team <repdata@sagebase.org>
}
\keyword{ tcga }
