% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer_info.R
\name{merge_data}
\alias{merge_data}
\title{Get info form maf and csv}
\usage{
merge_data(cancer, tcga_path, filter = FALSE)
}
\arguments{
\item{cancer}{The Chinese cancer name.}

\item{tcga_path}{TCGA data path contain maf and csv.}

\item{filter}{filter mutation?}
}
\value{
data.table
}
\description{
Get info form maf and csv
}
\examples{
\dontrun{
merge_data("胃癌", "data/path")
}
}
